# Conflicts:
#	common/src/main/java/org/terraform/biome/cavepopulators/DeepCavePopulator.java
#	common/src/main/java/org/terraform/biome/cavepopulators/FrozenCavePopulator.java
#	common/src/main/java/org/terraform/biome/cavepopulators/MossyCavePopulator.java
#	common/src/main/java/org/terraform/main/config/TConfigOption.java
#	common/src/main/java/org/terraform/structure/catacombs/CatacombsPathPopulator.java
#	common/src/main/java/org/terraform/structure/monument/MonumentDesign.java
#	common/src/main/java/org/terraform/structure/monument/MonumentPopulator.java
#	common/src/main/java/org/terraform/structure/pillager/mansion/MansionRoofHandler.java
#	common/src/main/java/org/terraform/structure/pillager/mansion/ground/MansionGroundLevelWarroomPopulator.java
#	common/src/main/java/org/terraform/structure/pillager/mansion/secondfloor/MansionSecondFloorBedroomPopulator.java
#	common/src/main/java/org/terraform/structure/pyramid/PyramidPopulator.java
#	common/src/main/java/org/terraform/structure/stronghold/StrongholdPathPopulator.java
#	common/src/main/java/org/terraform/structure/village/plains/forge/PlainsVillageForgeEntrancePiece.java
#	common/src/main/java/org/terraform/structure/village/plains/forge/PlainsVillageForgeJigsawBuilder.java
#	common/src/main/java/org/terraform/structure/village/plains/forge/PlainsVillageForgeRoofHandler.java
#	common/src/main/java/org/terraform/structure/village/plains/house/PlainsVillageEntrancePiece.java
#	common/src/main/java/org/terraform/structure/village/plains/house/PlainsVillageStandardPiece.java
#	common/src/main/java/org/terraform/structure/village/plains/temple/PlainsVillageTempleEntrancePiece.java
#	common/src/main/java/org/terraform/structure/villagehouse/animalfarm/AnimalFarmPopulator.java
#	common/src/main/java/org/terraform/tree/MushroomBuilder.java
#	common/src/main/java/org/terraform/tree/NewFractalTreeBuilder.java
#	common/src/main/java/org/terraform/tree/VanillaMushroomBuilder.java
#	common/src/main/java/org/terraform/utils/CylinderBuilder.java
#	common/src/main/java/org/terraform/utils/SphereBuilder.java
#	common/src/main/java/org/terraform/utils/blockdata/AgeableBuilder.java
#	common/src/main/java/org/terraform/utils/blockdata/BisectedBuilder.java

+++ common/src/main/java/org/terraform/biome/flat/DesertHandler.java
@@ -104,7 +104,7 @@ public class DesertHandler extends BiomeHandler {
 	public void spawnRibCage(@NotNull Random random, @NotNull SimpleBlock target) {
 		BlockFace direction = BlockUtils.getDirectBlockFace(random);
 		int spineLength = GenUtils.randInt(random, 10, 14);
-		float ribWidthRadius = GenUtils.randInt(random, 1, 2) + spineLength/2;
+		float ribWidthRadius = GenUtils.randInt(random, 1, 2) + (float) spineLength /2;
 		float ribHeightRadius = 0.7f*ribWidthRadius; //GenUtils.randInt(random, 6, 8);
 		//eqn -> ((y-ribHeight)/(ribHeight))^2 + ((x)/(ribWidth))^2 = 1
 		int interval = 2;
@@ -125,7 +125,7 @@ public class DesertHandler extends BiomeHandler {
 				ribSizeMultiplier -= 0.05f;
 			
 			if(segmentIndex % interval == 0 && segmentIndex > spineLength/6) {
-				for(float nHor = 1; nHor <= ribWidthRadius*ribSizeMultiplier; nHor+=0.01) {
+				for(float nHor = 1; nHor <= ribWidthRadius*ribSizeMultiplier; nHor+= 0.01F) {
 					
 					int[] multipliers = {-1};
 					if(nHor > ribWidthRadius*ribSizeMultiplier/3)
